backend/requirements.txt
Django>=5.1,<6.0
djangorestframework>=3.15
django-cors-headers>=4.3
django-filter>=24.0
psycopg[binary]>=3.2
python-decouple>=3.8
Pillow>=10.0
gunicorn>=22.0
whitenoise>=6.5
dj-database-url>=2.1
reportlab>=4.0
qrcode[pil]>=7.4
